fix signed mismatch when testing proper length of keyed hash.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-1-patches@9852 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/scapi.c
@@ -293,7 +293,7 @@ sc_generate_keyed_hash(const oid * authtype, size_t authtypelen,
     else {
         QUITFUN(SNMPERR_GENERR, sc_generate_keyed_hash_quit);
     }
-    if (buf_len != properlength) {
+    if ((int)buf_len != properlength) {
         QUITFUN(rval, sc_generate_keyed_hash_quit);
     }
     if ((int)*maclen > buf_len)
